structured the components
