+++ code/src/components/Main/TaskPage.js
@@ -1,0 +1,17 @@
+import React from "react";
+
+import { Header } from "../Header/Header";
+import { FormDrawer } from "../TaskForm/FormDrawer";
+import { TaskList } from "../TaskList/TaskList";
+import { Main } from "../../styledcomponents/StyledMain";
+
+
+export const TaskPage = () => {
+  return (
+    <Main>
+      <Header />
+      <FormDrawer />
+      <TaskList />
+    </Main>
+  )
+};
