+++ code/src/components/TaskForm/PickDate.js
@@ -2,23 +2,7 @@ import React from "react";
 import DatePicker from "react-datepicker";
 import "react-datepicker/dist/react-datepicker.css";
 
-import styled from "styled-components";
-
-const DateContainer = styled.div`
-  display: flex;
-  flex-direction: column;
-  margin-left: 30px;
-`;
-
-const Label = styled.label`
-  font-size: 15px;
-  font-weight: bold;
-  margin-bottom: 5px;
-  color: ${props => props.theme.textColor};
-  @media (min-width: 768px) {
-    font-size: 17px;
-  }
-`;
+import { DateContainer, Label } from "../../styledcomponents/StyledTaskForm";
 
 
 export const PickDate = ({ onChange, selected }) => {
