+++ code/src/styledcomponents/StyledTaskForm.js
@@ -1,5 +1,26 @@
 import styled from "styled-components";
 
+// For FormDrawer
+export const BottomContainer = styled.div`
+  display: flex;
+  justify-content: center;
+  flex-direction: row;
+  align-items: center;
+  position: fixed;
+  bottom: 0;
+  height: 75px;
+  width: 95%;
+  border-top: 2px solid ${props => props.theme.secondary};
+  background-color: ${props => props.theme.backgroundColor};
+  z-index: 1;
+  @media (min-width: 768px) {
+    height: 120px;
+  }
+  @media (min-width: 1025px) {
+    max-width: 900px;
+  }
+`;
+
 export const AddText = styled.p`
   font-size: 48px;
   margin: 0 0 6px 0;
@@ -16,6 +37,7 @@ export const DrawerContainer = styled.section`
   background-color: ${props => props.theme.drawer};
 `;
 
+// For AddTask
 export const Form = styled.form`
   display: flex;
   align-items: flex-start;
@@ -39,26 +61,6 @@ export const Input = styled.input`
   font-family: "Playfair Display";
 `;
 
-export const BottomContainer = styled.div`
-  display: flex;
-  justify-content: center;
-  flex-direction: row;
-  align-items: center;
-  position: fixed;
-  bottom: 0;
-  height: 75px;
-  width: 95%;
-  border-top: 2px solid ${props => props.theme.secondary};
-  background-color: ${props => props.theme.backgroundColor};
-  z-index: 1;
-  @media (min-width: 768px) {
-    height: 120px;
-  }
-  @media (min-width: 1025px) {
-    max-width: 900px;
-  }
-`;
-
 export const Options = styled.label`
   display: flex;
   flex-direction: column;
@@ -80,4 +82,21 @@ export const CategoryDate = styled.div`
   align-items: flex-end;
   width: 100%;
   margin: 0;
+`;
+
+// For PickDate
+export const DateContainer = styled.div`
+  display: flex;
+  flex-direction: column;
+  margin-left: 30px;
+`;
+
+export const Label = styled.label`
+  font-size: 15px;
+  font-weight: bold;
+  margin-bottom: 5px;
+  color: ${props => props.theme.textColor};
+  @media (min-width: 768px) {
+    font-size: 17px;
+  }
 `;
